requirements.txt
aiogram==3.0.0b7
pandas==2.0.3
python-dotenv==1.0.0
telethon==1.28.5
openpyxl==3.1.2